apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-multipart==0.0.20
